fix issues

+++ app/Http/Controllers/Admin/EditPageController.php
@@ -4,7 +4,9 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Models\EditPage;
+use App\Services\LinkTransformService;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 
 class EditPageController extends Controller
 {
@@ -14,9 +16,9 @@ class EditPageController extends Controller
     }
     public function edit(Request $request, $id) {
         $page_content = EditPage::query()->findOrFail($id);
-        return view('admin.pages.edit', compact('page_content'));
+        $jobs = json_decode(Storage::disk('local')->get('roles.json'), true);
+        return view('admin.pages.edit', compact('page_content', 'jobs'));
     }
-
     public function store(Request $request) {
         $data = $request->validate([
             'route_name' => ['required', 'string', 'unique:edit_pages,route_name'],
@@ -26,6 +28,10 @@ class EditPageController extends Controller
             'h1' => ['required', 'string'],
             'title' => ['required', 'string']
         ]);
+
+        foreach ($data as $key=>$value) {
+            $data[$key] = str_replace('</p>', '', str_replace('<p>', '', $value));
+        }
         $page = new EditPage($data);
         if ($page->save()) {
             return redirect()->route('admin.pages.index')->with('success', 'Page create successfully');
@@ -34,7 +40,8 @@ class EditPageController extends Controller
     }
 
     public function create() {
-        return view('admin.pages.create');
+        $jobs = json_decode(Storage::disk('local')->get('roles.json'), true);
+        return view('admin.pages.create', compact('jobs'));
     }
 
     public function update(Request $request) {
@@ -57,6 +64,9 @@ class EditPageController extends Controller
 
     public function delete(Request $request, $id) {
         $page_content = EditPage::query()->findOrFail($id);
+        if ($page_content->route_name === EditPage::DEFAULT_PAGE) {
+            return redirect()->back()->with('error', 'Default page cannot be deleted');
+        }
         $page_content->delete();
         return redirect()->back()->with('success', 'Page successfully deleted');
     }
